moved body parsing function to utils.js
